PDF-RAG Chatbot/backend/database.py
from unittest import loader
from dotenv import load_dotenv
from langchain_community.vectorstores import FAISS
from langchain_text_splitters import RecursiveCharacterTextSplitter
from langchain_community.document_loaders import PyPDFLoader
from langchain_community.embeddings import HuggingFaceEmbeddings

load_dotenv()

embeddings = HuggingFaceEmbeddings(
        model_name="sentence-transformers/all-MiniLM-L6-v2"
    )

def get_retriever(file_path):
    loader = PyPDFLoader(file_path)
    chunks = loader.load()
    text_splitter = RecursiveCharacterTextSplitter(chunk_size = 1000, chunk_overlap = 100)
    docs = text_splitter.split_documents(chunks)
    db = FAISS.from_documents(docs, embeddings)
    retriever = db.as_retriever(k=5)

    return retriever