PDF-RAG Chatbot/backend/requirements.txt
ipykernel>=6.29.5
langchain>=0.3.25
langgraph>=0.4.8
langchain-groq>=0.3.2
langchain-tavily>=0.2.3
langsmith>=0.3.45
python-dotenv>=1.1.0
langchain-google-genai
langchain-core
langchain-community
langchain-text-splitters
sentence-transformers
fastapi
uvicorn
streamlit
pypdf
PyPDF2
mcp
langchain_mcp_adapters
langgraph-cli[inmem]
chromadb
faiss-cpu
langchain-experimental
